sanity

+++ src/main/java/com/wylxbot/wylx/Database/Codecs/WylxStatsCodec.java
@@ -11,13 +11,13 @@ public class WylxStatsCodec implements Codec<WylxStats> {
     @Override
     public WylxStats decode(BsonReader reader, DecoderContext decoderContext) {
         long commandsProcessed = reader.readInt64("CommandsProcessed");
-        long averageCommandTime = reader.readInt64("AverageCommandTime");
+        double averageCommandTime = reader.readDouble("AverageCommandTime");
 
         long silentEventsProcessed = reader.readInt64("SilentEventsProcessed");
-        long averageSilentEventTime = reader.readInt64("AverageSilentEventTime");
+        double averageSilentEventTime = reader.readDouble("AverageSilentEventTime");
 
         long noOpsProcessed = reader.readInt64("NoOpsProcessed");
-        long averageNoOpTime = reader.readInt64("AverageNoOpTime");
+        double averageNoOpTime = reader.readDouble("AverageNoOpTime");
 
         return new WylxStats(commandsProcessed, averageCommandTime,
                 silentEventsProcessed, averageSilentEventTime,
@@ -27,13 +27,13 @@ public class WylxStatsCodec implements Codec<WylxStats> {
     @Override
     public void encode(BsonWriter writer, WylxStats value, EncoderContext encoderContext) {
         writer.writeInt64("CommandsProcessed", value.getCommandsProcessed());
-        writer.writeInt64("AverageCommandTime", value.getAverageCommandTime());
+        writer.writeDouble("AverageCommandTime", value.getAverageCommandTime());
 
         writer.writeInt64("SilentEventsProcessed", value.getSilentEventsProcessed());
-        writer.writeInt64("AverageSilentEventTime", value.getAverageSilentEventTime());
+        writer.writeDouble("AverageSilentEventTime", value.getAverageSilentEventTime());
 
         writer.writeInt64("NoOpsProcessed", value.getNoOpsProcessed());
-        writer.writeInt64("AverageNoOpTime", value.getAverageNoOpTime());
+        writer.writeDouble("AverageNoOpTime", value.getAverageNoOpTime());
     }
 
     @Override
